+++ packages/backtest/src/analyzers/Positions.ts
@@ -1,0 +1,12 @@
+import dayjs from 'dayjs'
+
+class PositionsAnalyzer {
+  name = 'positions'
+  data = []
+
+  end({ strategy }) {
+    this.data = strategy.broker.getPositions()
+  }
+}
+
+export default PositionsAnalyzer