+++ packages/backtest/src/analyzers/index.ts
@@ -1,1 +1,2 @@
 export { default as RetursAnalyzer } from './Returns'
+export { default as PositionsAnalyzer } from './Positions'