+++ packages/backtest/src/index.ts
@@ -1,2 +1,2 @@
-export * from './analyzers'
+export * as analyzers from './analyzers'
 export * from './core'