+++ packages/cli/src/api/cache.ts
@@ -3,6 +3,9 @@ import * as path from 'node:path'
 import { gzipSync, gunzipSync } from 'node:zlib'
 import clc from 'cli-color'
 import dayjs from 'dayjs'
+import duration from 'dayjs/plugin/duration'
+
+dayjs.extend(duration)
 
 export default (config) => {
   
@@ -46,7 +49,24 @@ export default (config) => {
   const get = (symbol: string, window: number, resolution: number = 1440, end?: string) => {
     const allData = getAll(symbol, resolution)
     const date = end ? dayjs(end) : dayjs()
-    const data = allData.filter((item) => item.date <= (resolution === 1440 ? date.endOf('D').valueOf() : date.valueOf()))
+    const fromDate = resolution === 1440 ? date.endOf('D') : date
+
+    // get data from specific date
+    const index = allData.findIndex(x => x.date <= fromDate.valueOf())
+    const data = allData.slice(index)
+
+    if (index === 0) {
+      const duration = dayjs.duration(data[0].date - data[1].date)
+      const fromDuration = dayjs.duration(fromDate.valueOf() - data[0].date)
+
+      if (fromDuration.asHours() > duration.asHours()) {
+        return []
+      }
+    }
+
+    if (index < 0) {
+      return []
+    }
 
     if (data.length < window) {
       return [];