+++ packages/backtest/CHANGELOG.md
@@ -1,5 +1,11 @@
 # zptrade-backtest
 
+## 1.0.7
+
+### Patch Changes
+
+- add analyzer option and fix cache
+
 ## 1.0.6
 
 ### Patch Changes